requirements.txt
click
apache-airflow[devel]>=1.10
pendulum
typing_extensions
pydantic